Remove duplicated code in trait selection
